feat: 기타

+++ order-core/src/main/java/dev/t1dmlgus/product/application/ProductInfo.java
@@ -48,7 +48,7 @@ public class ProductInfo {
             return ProductDetail.builder()
                     .productToken(product.getProductToken())
                     .productName(product.getName())
-                    .price(product.getPrice().getValue())
+                    .price((int) product.getPrice().getValue())
                     .productStatus(product.getStatus().getDescription())
                     .build();
         }
